chore: Add dynamic configuration

+++ spring-core/src/main/java/io/librevents/infrastructure/configuration/source/env/model/node/subscription/block/BlockSubscriptionConfigurationProperties.java
@@ -1,14 +1,16 @@
 package io.librevents.infrastructure.configuration.source.env.model.node.subscription.block;
 
+import java.math.BigInteger;
+
 import io.librevents.infrastructure.configuration.source.env.model.node.subscription.SubscriptionConfigurationProperties;
 import io.librevents.infrastructure.configuration.source.env.model.node.subscription.block.method.BlockSubscriptionMethodProperties;
 
 public record BlockSubscriptionConfigurationProperties(
         BlockSubscriptionMethodProperties method,
-        Integer initialBlock,
-        Integer confirmationBlocks,
-        Integer missingTxRetryBlocks,
-        Integer eventInvalidationBlockThreshold,
-        Integer replayBlockOffset,
-        Integer syncBlockLimit)
+        BigInteger initialBlock,
+        BigInteger confirmationBlocks,
+        BigInteger missingTxRetryBlocks,
+        BigInteger eventInvalidationBlockThreshold,
+        BigInteger replayBlockOffset,
+        BigInteger syncBlockLimit)
         implements SubscriptionConfigurationProperties {}
